requirements.txt
python-dotenv
PyPDF2
pypdf
langchain
ipykernel
supabase
langchain_community
tiktoken
langchain-openai
langchainhub
faiss-cpu
langgraph
langchain-anthropic
beautifulsoup4
duckduckgo-search
ollama
paramiko
pyaudio